requirements.txt
cython==0.20

numpy==1.17.5
pandas==0.25.3
python_daemon
Requests
smbus2
spidev
